update:: 아이템 추가 기능
