+++ ShoppingList/ShoppingList/Base.lproj/Main.storyboard
@@ -86,6 +86,9 @@
                                         <state key="normal" title="Add">
                                             <color key="titleColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                         </state>
+                                        <connections>
+                                            <action selector="addButtonClicked:" destination="BYZ-38-t0r" eventType="touchUpInside" id="cdh-he-P4Z"/>
+                                        </connections>
                                     </button>
                                 </subviews>
                             </stackView>
